+++ internal/user/imap.go
@@ -18,8 +18,10 @@
 package user
 
 import (
+	"bytes"
 	"context"
 	"fmt"
+	"net/mail"
 	"sync/atomic"
 	"time"
 
@@ -31,6 +33,7 @@ import (
 	"github.com/ProtonMail/proton-bridge/v3/internal/safe"
 	"github.com/ProtonMail/proton-bridge/v3/internal/vault"
 	"github.com/ProtonMail/proton-bridge/v3/pkg/message"
+	"github.com/ProtonMail/proton-bridge/v3/pkg/message/parser"
 	"github.com/bradenaw/juniper/stream"
 	"github.com/bradenaw/juniper/xslices"
 	"golang.org/x/exp/slices"
@@ -326,7 +329,7 @@ func (conn *imapConnector) RemoveMessagesFromMailbox(ctx context.Context, messag
 		return err
 	}
 
-	if mailboxID == proton.SpamLabel || mailboxID == proton.TrashLabel || mailboxID == proton.DraftsLabel {
+	if mailboxID == proton.TrashLabel || mailboxID == proton.DraftsLabel {
 		var metadata []proton.MessageMetadata
 
 		// There's currently no limit on how many IDs we can filter on,
@@ -437,20 +440,37 @@ func (conn *imapConnector) importMessage(
 
 	if err := safe.RLockRet(func() error {
 		return withAddrKR(conn.apiUser, conn.apiAddrs[conn.addrID], conn.vault.KeyPass(), func(_, addrKR *crypto.KeyRing) error {
-			res, err := stream.Collect(ctx, conn.client.ImportMessages(ctx, addrKR, 1, 1, []proton.ImportReq{{
-				Metadata: proton.ImportMetadata{
-					AddressID: conn.addrID,
-					LabelIDs:  labelIDs,
-					Unread:    proton.Bool(unread),
-					Flags:     flags,
-				},
-				Message: literal,
-			}}...))
-			if err != nil {
-				return fmt.Errorf("failed to import message: %w", err)
+			messageID := ""
+
+			if slices.Contains(labelIDs, proton.DraftsLabel) {
+				msg, err := conn.createDraft(ctx, literal, addrKR, conn.apiAddrs[conn.addrID])
+				if err != nil {
+					return fmt.Errorf("failed to create draft: %w", err)
+				}
+
+				// apply labels
+
+				messageID = msg.ID
+			} else {
+				res, err := stream.Collect(ctx, conn.client.ImportMessages(ctx, addrKR, 1, 1, []proton.ImportReq{{
+					Metadata: proton.ImportMetadata{
+						AddressID: conn.addrID,
+						LabelIDs:  labelIDs,
+						Unread:    proton.Bool(unread),
+						Flags:     flags,
+					},
+					Message: literal,
+				}}...))
+				if err != nil {
+					return fmt.Errorf("failed to import message: %w", err)
+				}
+
+				messageID = res[0].MessageID
 			}
 
-			if full, err = conn.client.GetFullMessage(ctx, res[0].MessageID); err != nil {
+			var err error
+
+			if full, err = conn.client.GetFullMessage(ctx, messageID); err != nil {
 				return fmt.Errorf("failed to fetch message: %w", err)
 			}
 
@@ -495,6 +515,63 @@ func toIMAPMessage(message proton.MessageMetadata) imap.Message {
 		Flags: flags,
 		Date:  date,
 	}
+}
+
+func (conn *imapConnector) createDraft(ctx context.Context, literal []byte, addrKR *crypto.KeyRing, sender proton.Address) (proton.Message, error) { //nolint:funlen
+	// Create a new message parser from the reader.
+	parser, err := parser.New(bytes.NewReader(literal))
+	if err != nil {
+		return proton.Message{}, fmt.Errorf("failed to create parser: %w", err)
+	}
+
+	message, err := message.ParseWithParser(parser)
+	if err != nil {
+		return proton.Message{}, fmt.Errorf("failed to parse message: %w", err)
+	}
+
+	decBody := string(message.PlainBody)
+	if message.RichBody != "" {
+		decBody = string(message.RichBody)
+	}
+
+	draft, err := conn.client.CreateDraft(ctx, addrKR, proton.CreateDraftReq{
+		Message: proton.DraftTemplate{
+			Subject:  message.Subject,
+			Body:     decBody,
+			MIMEType: message.MIMEType,
+
+			Sender:  &mail.Address{Name: sender.DisplayName, Address: sender.Email},
+			ToList:  message.ToList,
+			CCList:  message.CCList,
+			BCCList: message.BCCList,
+
+			ExternalID: message.ExternalID,
+		},
+	})
+
+	if err != nil {
+		return proton.Message{}, fmt.Errorf("failed to create draft: %w", err)
+	}
+
+	for _, att := range message.Attachments {
+		disposition := proton.AttachmentDisposition
+		if att.Disposition == "inline" && att.ContentID != "" {
+			disposition = proton.InlineDisposition
+		}
+
+		if _, err := conn.client.UploadAttachment(ctx, addrKR, proton.CreateAttachmentReq{
+			MessageID:   draft.ID,
+			Filename:    att.Name,
+			MIMEType:    rfc822.MIMEType(att.MIMEType),
+			Disposition: disposition,
+			ContentID:   att.ContentID,
+			Body:        att.Data,
+		}); err != nil {
+			return proton.Message{}, fmt.Errorf("failed to add attachment to draft: %w", err)
+		}
+	}
+
+	return draft, nil
 }
 
 func toIMAPMailbox(label proton.Label, flags, permFlags, attrs imap.FlagSet) imap.Mailbox {